A couple of small changes in how navigation works:
Settings:
o The settings dialog is now a modal dialog that slides in from the bottom. This works better than a push onto the navigation bar since the action is initiated from a button on the navigation bar instead of a table view cell.
o To apply the settings in the settings view controller the user now have to press the done button at the top left. This is more in line with Apple HiG for how to dismiss modal view controllers.
o The individual cells in the settings view controller now uses the «detail»-disclosure style instead of the «detail reveal»-disclosure style.
o The «about» button has been removed from the navigation bar and have been moved into the table as a cell instead. The reason behind this is that push from a button in a navigation bar feels and looks weird and since we are already in a modal view controller I did not want to present another modal view. Therefore we continue to push the about view controller onto the settings view controller, but from a table view cell instead.

Search
o  The search dialog is now a modal dialog that slides in from the bottom. This works better than a push onto the navigation controller since the action is initiated from a button on the navigation bar instead of a table view cell.