Refs #126 Start refresh of conferences and rely on the mechanism to auto-select latest conference if user tries to refresh sessions when no conference has been selected.
This fixes a lot of issues we previously had. For instance; Trying to refresh sessions if no conferences was loaded would previously not work. Users would have to understand that he had to go to settings and load conferences first. This situation was easy to get into if app was initially loaded with no internet connection.