Refs #126 Bigger refactoring: Please look out for regressions.
o Moved active conference responsibility from EMSAppDelegate to EMSRetriever.
o Split the session sync code into it´s own class. This class is a prototype that is used by EmsRetriever to sync a single session. The nice thing is that it can easily be cancelled and discarded in the cases where a user switch to another conference in the middle of a sync operation.