Refs #126 Removed the magic.
o EMSMainViewController now starts a sync when first loaded if there is no sessions found in search.
o Removed the sync code from EMSAppDelegate
o We no longer automatically start a sync of sessions when a new «active conference» is stored in the user defaults. Instead we do this in the unwindSettingsSegue again.