Solved the cell jumping issue that appeared when scrolling far down in the list and then adding/removing a session to the favorites.
The issues seems to be that the tableview uses the cell height estimate to do some calculations when re-configuring the cell. The error in estimate then seems to add up so the jump gets worse over time. Caching the cell-height as the cell appears when scrolling down and then using those values as estimate when the tableview asks for new estimates during cell reconfiguration seems to resolve the issue.